FIX: Regression in retrying introduced by 4b118378
The commit 4b118378 introduced a case of an "if-statement with an empty
body", causing the program to attempt retrying with both succeeded and
failed transfers (when --no-download was not used).  In doing so, the
program would also completely skip invoking the --exec command with the
succeeded transfers (when retried).

Affected versions:
  v0.7.12

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>