Add cc::input -- improve input parsing
The new implementation improves the input parsing:
* Support reading RFC2483 (making it possible to use with quvi 0.9+)
* Input URLs are now properly unescaped automagically

The command will read stdin by default. The input is expected to contain
URLs. The command arguments are expected to be either URLs or file
paths. If the input is read from either stdin or a file, the contents
are read as RFC2483.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>