Add --timestamp: Try to preserve modification time
Try to preserve the file modification time (as returned by the server,
if any). Using this option will cause the program to change the
modification time of the file to that of returned by the server.

NOTE: libcurl will parse this value from the returned "Last-Modified"
      HTTP header. This header may not always be present in which case
      the program will quietly ignore this option.

This is otherwise identical to Alexey's patch (thanks!) except that it
uses:
* Boost.Filesystem (last_write_time) to change the modification time
* the 'flags' of cclive's options

See also (for the original patch):
* http://www.mail-archive.com/cclive-devel@lists.sourceforge.net/msg00001.html

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>